Add support for xtl:endianness
This tests at runtime whether the target platform is little, big or mixed endian.
It also replace the configuration-time check that previously existed, just for
the sake of testing.

Note that C++20 proposes std::endian as well as a constexpr std::bit_cast, but
we're not there yet.

And even if a set of macro checks could replace that runtime check, it's much
more portable and maintainable to do it that way, esp. as both gcc and clang
actually fold that function.